refactor: 추상클래스를 인터페이스로 변환
