+++ src/main/java/org/example/movie/AmountDiscountPolicy.java
@@ -1,6 +1,6 @@
 package org.example.movie;
 
-public class AmountDiscountPolicy extends DiscountPolicy {
+public class AmountDiscountPolicy extends DefaultDiscountPolicy {
 
     private Money discountAmount;
