+++ src/main/java/org/example/movie/NonDiscountPolicy.java
@@ -1,0 +1,8 @@
+package org.example.movie;
+
+public class NonDiscountPolicy implements DiscountPolicy {
+    @Override
+    public Money calculateDiscountAmount(Screening screening) {
+        return Money.ZERO;
+    }
+}
