+++ src/main/java/org/example/movie/PercentDiscountPolicy.java
@@ -1,6 +1,6 @@
 package org.example.movie;
 
-public class PercentDiscountPolicy extends DiscountPolicy {
+public class PercentDiscountPolicy extends DefaultDiscountPolicy {
     private double percent;
 
     public PercentDiscountPolicy(double percent, DiscountCondition... conditions) {
